third_party/blink/web_tests/platform/generic/media/autoplay/console-warning-gesture-media-expected.txt

CONSOLE WARNING: <iframe gesture="media"> is not supported. Use <iframe allow="autoplay">, https://goo.gl/ximf56
